+++ CHANGELOG.md
@@ -6,4 +6,5 @@
   - update data
   - search data
   - create database if not exists
+  - create class object for Base, no need to define tablename